Add a compute_dtype field to GGMLTensor.

+++ invokeai/backend/model_manager/util/model_util.py
@@ -57,7 +57,7 @@ def read_checkpoint_meta(path: Union[str, Path], scan: bool = False) -> Dict[str
             if scan_result.infected_files != 0:
                 raise Exception(f'The model file "{path}" is potentially infected by malware. Aborting import.')
         if str(path).endswith(".gguf"):
-            checkpoint = gguf_sd_loader(Path(path))
+            checkpoint = gguf_sd_loader(Path(path), compute_dtype=torch.float32)
         else:
             checkpoint = torch.load(path, map_location=torch.device("meta"))
     return checkpoint
